+++ Fermion.cpp
@@ -2,10 +2,10 @@
 
 // Constructors
 // Parameterised
-Fermion::Fermion(double energy, double px, double py, double pz, float mass_in, float charge_in, bool anti_in):
+Fermion::Fermion(double energy, double px, double py, double pz, double mass_in, double charge_in, bool anti_in):
 Particle(energy, px, py, pz, spin, mass_in, charge_in, anti_in){}
 
-Fermion::Fermion(vector<double> energy_in, float mass_in, float charge_in, bool anti_in):
+Fermion::Fermion(vector<double> energy_in, double mass_in, double charge_in, bool anti_in):
 Particle(energy_in, spin, mass_in, charge_in, anti_in){}
 // Copy
 Fermion::Fermion(Fermion& fermion_in): Particle(fermion_in){}
