+++ Particle.cpp
@@ -2,7 +2,7 @@
 
 // Constructors
 // Parameterised
-Particle::Particle(double energy, double px, double py, double pz, float spin_in, float mass_in, float charge_in, bool anti_in)
+Particle::Particle(double energy, double px, double py, double pz, double spin_in, double mass_in, double charge_in, bool anti_in)
 {
   four_momentum = std::make_unique<FourMomentum>(FourMomentum(energy, px, py, pz));
   rest_mass = mass_in;
@@ -11,7 +11,7 @@ Particle::Particle(double energy, double px, double py, double pz, float spin_in
   spin = spin_in;
 }
 
-Particle::Particle(vector<double> momentum_in, float spin_in, float mass_in, float charge_in, bool anti_in)
+Particle::Particle(vector<double> momentum_in, double spin_in, double mass_in, double charge_in, bool anti_in)
 {
   four_momentum = std::make_unique<FourMomentum>(FourMomentum(momentum_in));
   rest_mass = mass_in;
@@ -75,10 +75,10 @@ Particle& Particle::operator=(Particle&& particle_in)
 }
 
 // Getters
-float Particle::get_rest_mass(){return rest_mass;}
+double Particle::get_rest_mass(){return rest_mass;}
 bool Particle::get_anti(){return is_anti;}
-float Particle::get_charge(){return charge;}
-float Particle::get_spin(){return spin;}
+double Particle::get_charge(){return charge;}
+double Particle::get_spin(){return spin;}
 FourMomentum Particle::get_four_momentum(){return *four_momentum;}
 
 // Setters
